Deprecate unused method arguments
I went through most of ActiveSupport with a static code analyzer, and
deprecated all method parameters/arguments whihc were not actually being
used in the method. I also went through Action Pack and Active Record
looking for "TODO" comments, to try and resolve.
