+++ actionpack/lib/abstract_controller/rendering.rb
@@ -122,8 +122,8 @@ module AbstractController
     # :api: private
     def _normalize_render(*args, &block)
       options = _normalize_args(*args, &block)
-      #TODO: remove defined? when we restore AP <=> AV dependency
-      if defined?(request) && request.variant.present?
+
+      if request.variant.present?
         options[:variant] = request.variant
       end
       _normalize_options(options)
