+++ activejob/lib/rails/generators/job/job_generator.rb
@@ -2,7 +2,7 @@ require 'rails/generators/named_base'
 
 module Rails # :nodoc:
   module Generators # :nodoc:
-    class JobGenerator < Rails::Generators::NamedBase # :nodoc:
+    class JobGenerator < NamedBase # :nodoc:
       desc 'This generator creates an active job file at app/jobs'
 
       class_option :queue, type: :string, default: 'default', desc: 'The queue name for the generated job'
